backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
aiosqlite
python-jose[cryptography]
python-dotenv
python-multipart
bcrypt